Refactor to build static site artifact
Update the ci configuration to produce a static site artifact instead of
building a Docker image. This makes for a much simpler, more efficient
deployment without.

This removes references in the previous build that produced a Docker
image, including the 'static' Dockerfile, an associated nginx config,
and a docker-compose configuration that was intended for local use.